refactor(UI-1269): update dashboard projects table header and row components
